refactor(bootstrap,application): rename InfrastructureException to BootstrapException and align exception packages
- Introduce BootstrapException as bootstrap-specific base
- Update ProfileConfigurationException to extend BootstrapException
- Align application exception package to application.error.exception
- Keep messageKey + args contract unchanged
